+++ src/components/basic_table/index.ts
@@ -1,0 +1,15 @@
+export {
+  EuiBasicTable,
+  EuiBasicTableProps,
+  EuiBasicTableColumn,
+} from './basic_table';
+export { EuiInMemoryTable, EuiInMemoryTableProps } from './in_memory_table';
+export {
+  EuiTableDataType,
+  EuiTableFooterProps,
+  EuiTableFieldDataColumnType,
+  EuiTableComputedColumnType,
+  EuiTableActionsColumnType,
+  EuiTableSelectionType,
+  EuiTableSortingType,
+} from './table_types';